tools/steamData/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
Pillow>=10.0.0
urllib3>=2.0.0
